Use size_t rather than unsigned for all cipher-related functions.

+++ aes-decrypt.c
@@ -338,7 +338,7 @@ _aes_decrypt_table =
 
 void
 aes_decrypt(const struct aes_ctx *ctx,
-	    unsigned length, uint8_t *dst,
+	    size_t length, uint8_t *dst,
 	    const uint8_t *src)
 {
   assert(!(length % AES_BLOCK_SIZE) );
